+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Proxy static assets for explorer instead of loading them directly from explorer.autometrics.dev. This url is also configurable so it is easier to use explorer hosted on a different URL
 
+- Add new endpoint `/api/functions` listing all autometrics'd functions in the current
+  working directory
+
 ## [0.5.0]
 
 - The Prometheus write endpoint is now enabled by default (#136)